Add follow view, profile view and getting mouse event
